LayoutTests/imported/w3c/web-platform-tests/html/semantics/interactive-elements/the-dialog-element/abspos-dialog-layout-expected.txt


FAIL showModal() should center in the viewport assert_approx_equals: expected 291 +/- 1 but got -500
FAIL Dialog should be recentered if showModal() is called after close() assert_approx_equals: expected 291 +/- 1 but got -1000
PASS Dialog should not recenter on relayout.
FAIL A tall dialog should be positioned at the top of the viewport. assert_equals: expected 0 but got -500
FAIL The dialog should be centered regardless of the presence of a horizontal scrollbar. assert_approx_equals: expected 0 +/- 1 but got -500
FAIL Centering should work when dialog is inside positioned containers. assert_approx_equals: expected 275 +/- 1 but got 330
PASS A centered dialog's position should survive becoming display:none temporarily.
FAIL Dialog should not still be centered when removed, and re-added to the document. assert_approx_equals: expected 285 +/- 1 but got 320
PASS Dialog's specified position should survive after close() and showModal().
FAIL Dialog should be recentered if showModal() is called after removing 'open'. assert_approx_equals: expected 0 +/- 1 but got -500
